backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.7
langchain-core==0.3.37
langchain-openai==0.2.8
langchain-google-genai==1.0.10
google-generativeai==0.8.3
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
PyPDF2==3.0.1
easyocr==1.7.1
opencv-python==4.10.0.84
pymongo>=4.9,<4.10
motor==3.6.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
numpy==1.26.4
sentence-transformers==3.0.1
rapidfuzz==3.10.1
openai==1.109.1
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil>=2.8.0
